model/pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.8",
    "datasets>=4.4.1",
    "dvc>=3.64.0",
    "dvc-ssh>=4.2.2",
    "emoji>=2.0.0",
    "mlflow>=3.7.0",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers[torch]>=4.57.3",
]
